WIP - refactor boilerplate into utils class
