+++ bert_trainer.py
@@ -9,6 +9,7 @@ from datetime import datetime
 
 import bert
 from bert import run_classifier
+import utils
 
 """
 Wrapper class to set up BERT model. Takes in Pandas DataFrame
@@ -22,14 +23,13 @@ class BERTTrainer():
 
     # These hyperparameters are copied from this colab notebook
     # (https://colab.sandbox.google.com/github/tensorflow/tpu/blob/master/tools/colab/bert_finetuning_with_cloud_tpus.ipynb)
-    def __init__(self, data, batch_size=32, max_seq_length=128,
+    def __init__(self, batch_size=32, max_seq_length=128,
             learning_rate=2e-5, num_train_epochs=3,
             warmup_proportion=0.1, save_checkpoints_every=500,
             save_summary_every=100, dropout_rate=0.5,
             is_training=True,
             bert_model_hub="https://tfhub.dev/google/bert_uncased_L-12_H-768_A-12/1",
             output_dir="bert_output"):
-        self.data = data
         self.batch_size = batch_size
         self.max_seq_length = max_seq_length
         self.learning_rate = learning_rate
@@ -43,124 +43,35 @@ class BERTTrainer():
         self.checkpoint_dir = output_dir if os.path.exists(output_dir) else None
         self.is_training = is_training
         print('Saving models to {}'.format(output_dir))
-        self._train_and_test_features_from_df()
-        self._create_estimator()
+        # self._train_and_test_features_from_df()
 
     '''
     Compute 80-20 train-valid split from data, and tokenize/pad sequences
     into required BERT input form.
     '''
 
-    def _train_and_test_features_from_df(self):
-        train, test = train_test_split(self.data, test_size=0.2, random_state=42)
-        print('Getting features for training and testing datasets')
+    """ model_fn_builder actually creates our model function using the passed parameters for num_labels, learning_rate, etc.
 
-        DATA_COLUMN = 'abstract'
-        LABEL_COLUMN = 'section'
-        self.label_list = list(self.data['section'].unique())
-
-        train_input_examples = train.apply(lambda x: bert.run_classifier.InputExample(
-            guid=None, # Globally unique ID for bookkeeping, unused in this example
-            text_a=x[DATA_COLUMN],
-            text_b=None,
-            label=x[LABEL_COLUMN]), axis=1)
-
-        test_input_examples = test.apply(lambda x: bert.run_classifier.InputExample(guid=None,
-            text_a=x[DATA_COLUMN],
-            text_b=None,
-            label=x[LABEL_COLUMN]), axis=1)
-
-        def _create_tokenizer_from_hub_module():
-            """Get the vocab file and casing info from the Hub module."""
-            with tf.Graph().as_default():
-                bert_module = hub.Module(self.bert_model_hub)
-                tokenization_info = bert_module(signature="tokenization_info", as_dict=True)
-                with tf.Session() as sess:
-                    vocab_file, do_lower_case = sess.run([tokenization_info["vocab_file"],
-                        tokenization_info["do_lower_case"]])
-
-                    return bert.tokenization.FullTokenizer(
-                            vocab_file=vocab_file, do_lower_case=do_lower_case)
-
-        self.tokenizer = _create_tokenizer_from_hub_module()
-
-        # Convert our train and test features to InputFeatures that BERT understands.
-        if self.is_training:
-            self.train_features = bert.run_classifier.convert_examples_to_features(
-                train_input_examples, self.label_list, self.max_seq_length, self.tokenizer)
-            self.test_features = bert.run_classifier.convert_examples_to_features(
-                test_input_examples, self.label_list, self.max_seq_length, self.tokenizer)
-
-    def _create_model(self, is_predicting, input_ids, input_mask, segment_ids, labels,
-            num_labels):
-        """ Creates a classification model.
-
-        Args:
-            is_predicting: Boolean to toggle training or testing mode.
-            (input_ids, input_mask, segment_ids, labels): Output of
-                bert.run_classifier.convert_examples_to_features
-            num_labels: Number of classes to classify.
-        Returns:
-            If training, loss and prediction tensors, otherwise just prediction tensors.
-        """
-
-        bert_module = hub.Module(
-                self.bert_model_hub,
-                trainable=True)
-        bert_inputs = dict(
-                input_ids=input_ids,
-                input_mask=input_mask,
-                segment_ids=segment_ids)
-        bert_outputs = bert_module(
-                inputs=bert_inputs,
-                signature="tokens",
-                as_dict=True)
-
-        # Use "pooled_output" for classification tasks on an entire sentence.
-        # Use "sequence_outputs" for token-level output.
-        output_layer = bert_outputs["pooled_output"]
-
-        hidden_size = output_layer.shape[-1].value
-
-        # Dense layer that outputs classes after BERT layer.
-        output_weights = tf.get_variable(
-                "output_weights", [num_labels, hidden_size],
-                initializer=tf.truncated_normal_initializer(stddev=0.02))
-
-        print(output_layer.shape, output_weights.shape)
-
-        output_bias = tf.get_variable(
-                "output_bias", [num_labels], initializer=tf.zeros_initializer())
-
-        with tf.variable_scope("loss"):
-
-            # Dropout helps prevent overfitting during training
-            output_layer = tf.nn.dropout(output_layer, keep_prob=self.dropout_rate, seed=42)
-
-            logits = tf.matmul(output_layer, output_weights, transpose_b=True)
-            logits = tf.add(logits, output_bias)
-            log_probs = tf.nn.log_softmax(logits, axis=-1)
-
-            # Convert labels into one-hot encoding
-            one_hot_labels = tf.one_hot(labels, depth=num_labels, dtype=tf.float32)
-
-            predicted_labels = tf.squeeze(tf.argmax(log_probs, axis=-1, output_type=tf.int32))
-            # If we're predicting, we want predicted labels and the probabiltiies.
-            if is_predicting:
-                return (predicted_labels, log_probs)
-
-            # If we're train/eval, compute loss between predicted and actual label
-            per_example_loss = -tf.reduce_sum(one_hot_labels * log_probs, axis=-1)
-            loss = tf.reduce_mean(per_example_loss)
-            return (loss, predicted_labels, log_probs)
-
-    # model_fn_builder actually creates our model function
-    # using the passed parameters for num_labels, learning_rate, etc.
+    Arguments:
+        num_labels: The number of classes in our classification model.
+        learning_rate: Float between 0 and 1 for learning rate.
+        num_train_steps: An integer for the number of training steps to run.
+        num_warmup_steps: An integer for the number of warmup steps to run before training.
+    """
     def _model_fn_builder(self, num_labels, learning_rate, num_train_steps,
             num_warmup_steps):
         """Returns `model_fn` closure for TPUEstimator."""
         def model_fn(features, labels, mode, params):  # pylint: disable=unused-argument
-            """The `model_fn` for TPUEstimator."""
+            """The `model_fn` for TPUEstimator.
+
+            Arguments:
+                features: list of features outputted by convert_examples_to_features
+                labels: unused variable, required in model_fn signature.
+                mode: are we training or testing?
+                params: hyperparameters for the training process
+            Returns:
+                A tf.EstimatorSpec used internally by the estimator instance.
+            """
 
             input_ids = features["input_ids"]
             input_mask = features["input_mask"]
@@ -172,8 +83,10 @@ class BERTTrainer():
             # TRAIN and EVAL
             if not is_predicting:
 
-                (loss, predicted_labels, log_probs) = self._create_model(
-                        is_predicting, input_ids, input_mask, segment_ids, label_ids, num_labels)
+                print(self.bert_model_hub)
+                (loss, predicted_labels, log_probs) = utils.create_model(
+                        self.bert_model_hub, is_predicting, input_ids, input_mask, segment_ids, label_ids,
+                        num_labels, self.dropout_rate)
 
                 train_op = bert.optimization.create_optimizer(
                         loss, learning_rate, num_train_steps, num_warmup_steps, use_tpu=False)
@@ -196,7 +109,7 @@ class BERTTrainer():
                             loss=loss,
                             eval_metric_ops=eval_metrics)
             else:
-                (predicted_labels, log_probs) = self._create_model(
+                (predicted_labels, log_probs) = utils.create_model(
                         is_predicting, input_ids, input_mask, segment_ids, label_ids, num_labels)
 
                 predictions = {
@@ -235,7 +148,10 @@ class BERTTrainer():
                 params={"batch_size": self.batch_size})
 
     # Create an input function for training. drop_remainder = True for using TPUs.
-    def train_model(self):
+    def train(self, train_features, label_list):
+        self.train_features = train_features
+        self.label_list = label_list
+        self._create_estimator()
         train_input_fn = bert.run_classifier.input_fn_builder(
                 features=self.train_features,
                 seq_length=self.max_seq_length,
